Made engine a separate library, built on top of the core Sierra API, prepared an architecture for handling both Runtime and Editor builds, improved template usages

+++ Sierra/src/Rendering/Platform/Vulkan/Platform/Linux/VulkanLinuxSurface.cpp
@@ -23,7 +23,7 @@ namespace Sierra
         // Create surface
         VkSurfaceKHR surface;
         const VkResult result = instance.GetFunctionTable().vkCreateXlibSurfaceKHR(instance.GetVulkanInstance(), &surfaceCreateInfo, nullptr, &surface);
-        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create Linux surface for window [{0}]! Error code: {1}.", window->GetTitle(), result);
+        SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create Linux surface for window [{0}]! Error code: {1}.", window->GetTitle(), static_cast<int32>(result));
 
         return surface;
     }
