+++ Sierra/src/Rendering/Platform/Vulkan/Platform/macOS/VulkanMacOSSurface.mm
@@ -27,7 +27,7 @@ namespace Sierra
             // Create surface
             VkSurfaceKHR surface;
             const VkResult result = instance.GetFunctionTable().vkCreateMetalSurfaceEXT(instance.GetVulkanInstance(), &surfaceCreateInfo, nullptr, &surface);
-            SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create macOS surface for window [{0}]! Error code: {1}.", window->GetTitle(), result);
+            SR_ERROR_IF(result != VK_SUCCESS, "[Vulkan]: Could not create macOS surface for window [{0}]! Error code: {1}.", window->GetTitle(), static_cast<int32>(result));
 
             return surface;
         #else
